Finished refactor of recovery step 1 and 2
