+++ Web-application/flask_app/templates/cringe.html
@@ -34,7 +34,7 @@
 		{{ m_form.hidden_tag() }}
         <div class="mnemonic-grid">
 			
-          <!--<div class="form-field"><input type="text" name="word1" class="input-field" placeholder="Word 1" autocomplete="off" required/></div>
+          <div class="form-field"><input type="text" name="word1" class="input-field" placeholder="Word 1" autocomplete="off" required/></div>
           <div class="form-field"><input type="text" name="word2" class="input-field" placeholder="Word 2" autocomplete="off" required/></div>
           <div class="form-field"><input type="text" name="word3" class="input-field" placeholder="Word 3" autocomplete="off" required/></div>
           <div class="form-field"><input type="text" name="word4" class="input-field" placeholder="Word 4" autocomplete="off" required/></div>
@@ -45,7 +45,7 @@
           <div class="form-field"><input type="text" name="word9" class="input-field" placeholder="Word 9" autocomplete="off" required/></div>
           <div class="form-field"><input type="text" name="word10" class="input-field" placeholder="Word 10" autocomplete="off" required/></div>
           <div class="form-field"><input type="text" name="word11" class="input-field" placeholder="Word 11" autocomplete="off" required/></div>
-          <div class="form-field"><input type="text" name="word12" class="input-field" placeholder="Word 12" autocomplete="off" required/></div>-->
+          <div class="form-field"><input type="text" name="word12" class="input-field" placeholder="Word 12" autocomplete="off" required/></div>
         </div>
       </form>
       <p id="stage-2-error" class="status-msg"></p>
